first registered user

+++ src/utils/cloudinary.js
@@ -3,10 +3,10 @@ import fs from "fs";
 
 // fs is the node js library for file system handling 
 
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
 // 
@@ -16,20 +16,29 @@ const Uploadoncloud = async (localfilepath) => {
         if (!localfilepath) {
             return null;
         }
-        else
-        {
+        else {
             // unpload the locally saved file to the cloud and get its URL 
 
-            const response = await cloudinary.uploader.upload(localfilepath,{
+            const response = await cloudinary.uploader.upload(localfilepath, {
                 resource_type: "auto"
             });
+
+            console.log("file uploaded on cloud", response.url);
+            if (fs.existsSync(localfilepath)) {
+                fs.unlinkSync(localfilepath);
+            }
+            return response;
         }
-        console.log("file uploaded on cloud",response.url);
-        fs.unlinkSync(localfilepath); // delete from the local storage
-        return response;
+
     } catch (error) {
-        fs.unlinkSync(localfilepath); // remove the locally saved temporary file incase upload opertion fail
+        console.log("Cloudinary error:", error);
+
+        if (fs.existsSync(localfilepath)) {
+            fs.unlinkSync(localfilepath);
+        }
+
+        return null;
     }
 }
 
-export {Uploadoncloud}
+export { Uploadoncloud }
